AutoarExtract: refactor destination decision process
Previously, the destination of an extraction could not be modified by the user
application once the AutoarExtract object was created. This leads to a big lack
of flexibility, because there are many cases when the destination depends on the
contents of the archive, unknown at object creation time.

In order to fix this, when the destination is notified to the user, also notify
the list of files in the archive and enable the user to set a new destination.

https://bugzilla.gnome.org/show_bug.cgi?id=768645