AutoarExtract: add signal for name conflicts
Extraction was previously done with overwrite as the default option for existing
files. A better solution would be to detect conflicts and notify the user
application, offering the possibility to change the destination of the extracted
file.

In order to fix this, add a signal for notifying conflicts. This signal gets
emitted before actually extracting a file.

https://bugzilla.gnome.org/show_bug.cgi?id=768645