configure: Ensure libarchive version is high enough in fallback check
The previous version of fallback check uses archive_read_new, which is
incorrect because it will wrongly accept an old version of libarchive
and fail at link time instead of configure time. This commit fixes it
by replacing archive_read_new with archive_entry_is_encrypted. This
function is added in libarchive 3.2.0 and systems with lower versions
will fail the check.

https://bugzilla.gnome.org/show_bug.cgi?id=770735