AutoarExtractor: treat missing case in destination suggestion process
The case of archives having the name of the root element with the format
extension (for example 'sample.odt' in 'sample.odt.zip') was not treated. This
lead to unnecessary folders being created. In order to fix this, check if the
archive's name without extension is the same as the root element's name.

https://bugzilla.gnome.org/show_bug.cgi?id=773520