release: Bump both package version and API version to 0.2
We used to have two API versions in this module. 0.1 was used in gir,
but 0 was used in include, lib, pc, vapi. To simplify versioning, both
are bumped to 0.2, which is the same as the package version of the
latest release that causes API break.

LIBGNOME_AUTOAR_API_VERSION autoconf output variable is renamed to
API_VERSION to make the variable name easier to read and type. A new
output variable, API_VERSION_U, is added for using with gir and vapi
because they require the dot in the version string to be replaced by
an underscore.

Since automake rejects Makefiles that do macro expansion before the
equal sign in macro definition lines, all references to API_VERSION
and API_VERSION_U in Makefiles use autoconf output variable instead
of macro expansion.