tests: Move programs not related to testing to a separate directory
There are several programs used for development and manual testing
put in the 'tests' directory. This is confusing for new developers
because the 'tests' directory is usually used to store automated test
programs and test cases. These programs are now put in the new 'tools'
directory, and their 'test-' prefix are also changed.