AutoarExtractor: Rename 'decide-destination' to make it synchronous
This commit is similar to commit 4ac0fd4, which was a temporary fix made
for gnome-autoar 0.1 releases. The new name of this synchronous signal
is 'change-destination', which is clearer because it mentions the
ability of the signal to change the destination,

The name 'decide-destination' still exists in the source code, which is
reserved for a new asynchronous signal that can report the destination
but cannot change it.

https://bugzilla.gnome.org/show_bug.cgi?id=770437#c9