mime-types: use g_content_type_is_a for mime support
We were using a plain text matching, which is less than ideal and prone
to error.

One clear practical issue is that rar was reported as not supported,
since the supertype is vnd.rar, not any alias like x-rar, etc.

This patch fix autoar to make sure we check for the supertypes for
reporting whether the types are supported or not.

https://bugzilla.gnome.org/show_bug.cgi?id=779431