mime-types: drop rar support
Unfortunately rar is not well supported by libarchive, so for now let's
drop its support so we don't expect users of autoar to extract files
that are corrupted.

The upstream documentation of what rar parts are supported or not are
not clear, so far I found:
https://github.com/libarchive/libarchive/issues/373
https://github.com/vitasdk/packages/pull/4#issuecomment-272375907
https://github.com/libarchive/libarchive/wiki/LibarchiveFormats it says
read only for rar, but that is should work for any rar. However seems
rar5 is not supported.

This patch drops support for rar, and in case of some clients like
Nautilus will use file-roller for rar until libarchive fixes.

https://bugzilla.gnome.org/show_bug.cgi?id=779437