extractor: Fix default action for conflicts
The `AUTOAR_CONFLICT_OVERWRITE` is set as default value for the action
variable when conflict occured. However, `g_signal_emit` clears that
variable to `0` if the signal is unhandled. But `0` is currently mapped
to `AUTOAR_CONFLICT_SKIP`. So the code is a little bit confusing. I think
that overwrite is the right thing in most cases and also this is the
default behavior of `tar` as an archive may contain several versions of
some file and the last one is the newest. However, gnome-autoar allows
extraction in the non-empty folders and has conflict API, so it would be
really safer to use the skip action by default. Let's add the
`AUTOAR_CONFLICT_UNHANDLED` action for better control and use the
`AUTOAR_CONFLICT_SKIP` action by default.