extractor: Do not allow prefix modifications when output-is-dest
If the `output-is-dest` property is `TRUE`, the `prefix` is not cleared and is
passed to `decide-destination` signal. This looks unexpected because it allows
to change even the prefix which doesn't match archive name, which is not allowed
even if `output-is-dest` is `FALSE`. I am conviced that it should not be allowed
to change the `prefix` at all in this case. Let's clear the `prefix` variable to
avoid that.