tests: Change test-one-file-error-file-over-directory test
To be honest, it is not really clear to me what is purpose of this test.
As per the name, it should verify that error is returned when overwriting
file over directory. However, I think that it is totally fine to overwrite
empty directory. Anyway, the overwrite action is not explicitely chosen,
so the skip action is used instead. Consequently, the test verifies that
`error` is not set. So it looks to me that the test is tottaly wrong. Let's
modify and rename the test, so it really checks that error is returned when
somebody tries to overwrite non-empty directory.