compressor: Fix hardlink detection for remote files
Currently, all files are considered as hardlinks when compressing files provided
by GVfs (e.g. SFTP) to TAR format, so the file content is lost for those files.
This is because the `inode`, `device`, and/or `nlink` attributes are not set for
those files. Let's do not call `archive_entry_linkify` in this case to not create
broken archives.

Fixes: https://gitlab.gnome.org/GNOME/gnome-autoar/-/issues/25