compressor: Fix hardlink handling for new cpio format
The new cpio format writes the file content for hardlinks the last
time an inode is seen. To achieve this, the archive entries are
internally stored over `archive_entry_linkify` and written later.
In that case, `archive_entry_linkify` takes ownership of the
`archive_entry` struct and set the pointer to `NULL`. However,
gnome-autoar keeps using the original entry, which leads to weird
errors. Also, after all archive entries are written,
`archive_entry_linkify` has to be called in a loop to write all
deffered entries. However, it is called only once currently, which
can cause that some of the hardlinks are not written to the
archive. Let's correctly handle the entry ownership and call
`archive_entry_linkify` to fix hardlink handling for the new cpio
format.

Relates: https://gitlab.gnome.org/GNOME/gnome-autoar/-/issues/25