compressor: Add encryption support
`AutoarExtractor` supports extraction of encrypted archives. It would be nice
to have support for encrypted archives also in `AutoarCompressor`. Though the
libarchive library doesn't provide a much. It supports encryption only for the
ZIP format. There is a choice between `zipcrypt`, `aes128` and `aes256`. Given
the phact that `zipcrypt` is insecure and not recommended, I think it is fine
to hardcode `aes256`. It is probably a bit slower then `aes128`, but more secure.
So the only thing which needs to be set is a passphrase. Let's add
`autoar_compressor_set_passphrase` to allow creating encrypted archives.

Relates: https://gitlab.gnome.org/GNOME/nautilus/-/issues/822