Add missing initialization for g_auto* variables
When building Fedora package, I see many "may be used uninitialzed"
warnings. I am not sure why but I don't see those warnings locally even
when `-Wmaybe-uninitialized` is used. Anyway, this is because the
`g_autofree` and `g_autoptr` variables are not explicitely initialized
as mentioned in the documentation. Let's initialize them to get rid of
those warnings.