extractor: Add autoar_extractor_set_passphrase function
To manage passphrase-protected files, the extractor sends the
`request-passphrase` signal and expects to receive the passphrase as
the signal's return value. That's too complicated for some use cases
and is not possible with asynchronous API currently. Let's add an
extra method that allows to set the passphrase.

Relates: https://gitlab.gnome.org/GNOME/gnome-autoar/-/issues/42