tests: Skip file size check for directories
With the latest GLib version, the test suite fails due to the
`GLib-GIO-FATAL-CRITICAL: GFileInfo created without standard::size`
critical error. This is because the `g_file_info_get_size` function
is called, but the `G_FILE_ATTRIBUTE_STANDARD_SIZE` was not queried.
This happens only for directories, where the file size check doesn't
make much sense anyway. Let's skip the file size check for directories
to get rid of the critical error resp. to fix the test suite.

Fixes: https://gitlab.gnome.org/GNOME/gnome-autoar/-/issues/44